+++ src/BaGet.Core/IPackageService.cs
@@ -1,115 +1,56 @@
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using System.Threading;
 using System.Threading.Tasks;
 using NuGet.Versioning;
 
 namespace BaGet.Core
 {
-    /// <summary>
-    /// The "source of truth" for packages' state. Packages' content
-    /// are stored by the <see cref="IPackageStorageService"/>.
-    /// </summary>
     public interface IPackageService
     {
         /// <summary>
-        /// Attempt to add a new package to the database.
+        /// Attempt to find a package's versions using mirroring. This will merge
+        /// results from the configured upstream source with the locally indexed packages.
         /// </summary>
-        /// <param name="package">The package to add to the database.</param>
-        /// <param name="cancellationToken">A token to cancel the task.</param>
-        /// <returns>The result of attempting to add the package to the database.</returns>
-        Task<PackageAddResult> AddAsync(Package package, CancellationToken cancellationToken);
+        /// <param name="id">The package's id to lookup</param>
+        /// <param name="cancellationToken">The token to cancel the lookup</param>
+        /// <returns>
+        /// The package's versions, or an empty list if the package cannot be found.
+        /// This includes unlisted versions.
+        /// </returns>
+        Task<IReadOnlyList<NuGetVersion>> FindPackageVersionsAsync(string id, CancellationToken cancellationToken);
 
         /// <summary>
-        /// Attempt to find a package with the given id and version.
+        /// Attempt to find a package's metadata using mirroring. This will merge
+        /// results from the configured upstream source with the locally indexed packages.
         /// </summary>
-        /// <param name="id">The package's id.</param>
-        /// <param name="version">The package's version.</param>
-        /// <param name="includeUnlisted">Whether unlisted results should be included.</param>
-        /// <param name="cancellationToken">A token to cancel the task.</param>
-        /// <returns>The package found, or null.</returns>
-        Task<Package> FindOrNullAsync(
-            string id,
-            NuGetVersion version,
-            bool includeUnlisted,
-            CancellationToken cancellationToken);
+        /// <param name="id">The package's id to lookup</param>
+        /// <param name="cancellationToken">The token to cancel the lookup</param>
+        /// <returns>
+        /// The metadata for all versions of a package, including unlisted versions.
+        /// Returns an empty list if the package cannot be found.
+        /// </returns>
+        Task<IReadOnlyList<Package>> FindPackagesAsync(string id, CancellationToken cancellationToken);
 
         /// <summary>
-        /// Attempt to find all packages with a given id.
+        /// Attempt to find a package's metadata using mirroring. This will merge
+        /// results from the configured upstream source with the locally indexed packages.
         /// </summary>
-        /// <param name="id">The packages' id.</param>
-        /// <param name="includeUnlisted">Whether unlisted results should be included.</param>
-        /// <param name="cancellationToken">A token to cancel the task.</param>
-        /// <returns>The packages found. Always non-null.</returns>
-        Task<IReadOnlyList<Package>> FindAsync(string id, bool includeUnlisted, CancellationToken cancellationToken);
+        /// <param name="id">The package's id to lookup</param>
+        /// <param name="version">The package's version to lookup</param>
+        /// <param name="cancellationToken">The token to cancel the lookup</param>
+        /// <returns>
+        /// The metadata for single version of a package.
+        /// Returns null if the package does not exist.
+        /// </returns>
+        Task<Package> FindPackageOrNullAsync(string id, NuGetVersion version, CancellationToken cancellationToken);
 
         /// <summary>
-        /// Determine whether a package exists in the database (even if the package is unlisted).
-        /// </summary>
-        /// <param name="id">The package id to search.</param>
-        /// <param name="cancellationToken">A token to cancel the task.</param>
-        /// <returns>Whether the package exists in the database.</returns>
-        Task<bool> ExistsAsync(string id, CancellationToken cancellationToken);
-
-        /// <summary>
-        /// Determine whether a package exists in the database (even if the package is unlisted).
+        /// Determine whether a package exists locally or on the upstream source.
         /// </summary>
         /// <param name="id">The package id to search.</param>
         /// <param name="version">The package version to search.</param>
         /// <param name="cancellationToken">A token to cancel the task.</param>
         /// <returns>Whether the package exists in the database.</returns>
         Task<bool> ExistsAsync(string id, NuGetVersion version, CancellationToken cancellationToken);
-
-        /// <summary>
-        /// Unlist a package, making it undiscoverable.
-        /// </summary>
-        /// <param name="id">The id of the package to unlist.</param>
-        /// <param name="version">The version of the package to unlist.</param>
-        /// <param name="cancellationToken">A token to cancel the task.</param>
-        /// <returns>False if the package does not exist.</returns>
-        Task<bool> UnlistPackageAsync(string id, NuGetVersion version, CancellationToken cancellationToken);
-
-        /// <summary>
-        /// Relist a package, making it discoverable.
-        /// </summary>
-        /// <param name="id">The id of the package to relist.</param>
-        /// <param name="version">The version of the package to relist.</param>
-        /// <param name="cancellationToken">A token to cancel the task.</param>
-        /// <returns>False if the package does not exist.</returns>
-        Task<bool> RelistPackageAsync(string id, NuGetVersion version, CancellationToken cancellationToken);
-
-        /// <summary>
-        /// Increment a package's download count.
-        /// </summary>
-        /// <param name="id">The id of the package to update.</param>
-        /// <param name="version">The id of the package to update.</param>
-        /// <param name="cancellationToken">A token to cancel the task.</param>
-        /// <returns>False if the package does not exist.</returns>
-        Task<bool> AddDownloadAsync(string id, NuGetVersion version, CancellationToken cancellationToken);
-
-        /// <summary>
-        /// Completely remove the package from the database.
-        /// </summary>
-        /// <param name="id">The id of the package to remove.</param>
-        /// <param name="version">The version of the pacakge to remove.</param>
-        /// <param name="cancellationToken">A token to cancel the task.</param>
-        /// <returns>False if the package doesn't exist.</returns>
-        Task<bool> HardDeletePackageAsync(string id, NuGetVersion version, CancellationToken cancellationToken);
-    }
-
-    /// <summary>
-    /// The result of attempting to add the package to the database.
-    /// See <see cref="IPackageService.AddAsync(Package, CancellationToken)"/>
-    /// </summary>
-    public enum PackageAddResult
-    {
-        /// <summary>
-        /// Failed to add the package as it already exists.
-        /// </summary>
-        PackageAlreadyExists,
-
-        /// <summary>
-        /// The package was added successfully.
-        /// </summary>
-        Success
     }
 }
