+++ src/BaGet.Core/Metadata/DefaultPackageMetadataService.cs
@@ -10,14 +10,14 @@ namespace BaGet.Core
     /// <inheritdoc />
     public class DefaultPackageMetadataService : IPackageMetadataService
     {
-        private readonly IMirrorService _mirror;
+        private readonly IPackageService _packages;
         private readonly RegistrationBuilder _builder;
 
         public DefaultPackageMetadataService(
-            IMirrorService mirror,
+            IPackageService packages,
             RegistrationBuilder builder)
         {
-            _mirror = mirror ?? throw new ArgumentNullException(nameof(mirror));
+            _packages = packages ?? throw new ArgumentNullException(nameof(packages));
             _builder = builder ?? throw new ArgumentNullException(nameof(builder));
         }
 
@@ -25,7 +25,7 @@ namespace BaGet.Core
             string packageId,
             CancellationToken cancellationToken = default)
         {
-            var packages = await _mirror.FindPackagesAsync(packageId, cancellationToken);
+            var packages = await _packages.FindPackagesAsync(packageId, cancellationToken);
             if (!packages.Any())
             {
                 return null;
@@ -42,7 +42,7 @@ namespace BaGet.Core
             NuGetVersion version,
             CancellationToken cancellationToken = default)
         {
-            var package = await _mirror.FindPackageOrNullAsync(id, version, cancellationToken);
+            var package = await _packages.FindPackageOrNullAsync(id, version, cancellationToken);
             if (package == null)
             {
                 return null;
