+++ src/BaGet.Core/Upstream/DownloadsImporter.cs
