+++ src/BaGet.Core/Upstream/IPackageDownloadsSource.cs
